OCR/requirements.txt
Flask
requests
google-generativeai
Werkzeug